+++ contracts/PreIcoStrategy.sol
@@ -10,13 +10,15 @@ contract  PreIcoStrategy is PricingStrategy {
     uint256 public bonus;
     uint256 public maxCap;
 
-    //events for testing
+    /*
+    * @dev event after counting tokens of a investor
+    */
     event TokenForPreIcoInvestor(uint256 _token, uint256 _tokenAndBonus, uint256 bonusRate);
 
     /*
     * @dev Constructor
     */
-    function PreIcoStrategy(uint256 _bonus, uint _maxCap, uint256 _rate) public {
+    function PreIcoStrategy(uint256 _bonus, uint256 _maxCap, uint256 _rate) public {
         bonus = _bonus;
         maxCap = _maxCap;
         rate = _rate;
@@ -27,7 +29,6 @@ contract  PreIcoStrategy is PricingStrategy {
     * @dev Count number of tokens with bonuses
     * @return uint256 Return number of tokens for an investor
     */
-
     function countTokens(uint256 _value) public onlyCrowdsale returns (uint256 tokensAndBonus) {
         uint256 etherInWei = 1e18;
         uint256 val = _value.mul(etherInWei);
@@ -56,17 +57,23 @@ contract  PreIcoStrategy is PricingStrategy {
         return (maxCap - totalSoldTokens) >= requiredTokens;
     }
 
+    /*
+    * @dev Calculate unsold tokens
+    */
     function getUnSoldTokens() public constant returns (uint256) {
         return maxCap - totalSoldTokens;
     }
 
+    /*
+    * @dev Check available tokens for sale
+    */
     function isNoEmptyTranches() public constant returns(bool) {
         uint256 availableTokens = maxCap - totalSoldTokens;
         return availableTokens > 0 && now <= endTime;
     }
 
     /*
-    * @dev Summing sold of tokens
+    * @dev Summing sold tokens
     */
     function soldInTranche(uint256 tokens) internal {
         totalSoldTokens = totalSoldTokens.add(tokens);
